pyproject.toml
[tool.poetry]
name = "up-book"
version = "0.1.0"
description = "A Book Upload Web App"
authors = ["Feyaaz Chishty"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^2.2.2"
black = "^23.1.0"
isort = "^5.12.0"
psycopg2-binary = "^2.9.5"
loguru = "^0.6.0"
boto3 = "^1.26.64"
flask-session = "^0.4.0"
pandas = "^1.5.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
